requirements.txt
# Core
tiktoken>=0.5.0

# LLM Providers
google-genai>=1.0.0
groq>=0.4.0
openai>=1.0.0
anthropic>=0.18.0
together>=1.0.0

# For experiments & analysis
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0

# For loading .env files
python-dotenv>=1.0.0
